refactor: refatora projeto removendo arquivos, comentários e criando pacotes coerentes
