+++ src/main/java/equipe25/churninsight_backend/model/explicabilidade/ExplicabilidadeEnum.java
@@ -1,4 +1,4 @@
-package equipe25.churninsight_backend.model.previsao;
+package equipe25.churninsight_backend.model.explicabilidade;
 
 public enum ExplicabilidadeEnum {
